Move LexicalHorizontalRule to lexical/react
