+++ packages/lexical-react/src/LexicalHorizontalRuleNode.js
